+++ src/main/java/ajou/mse/dimensionguard/repository/PlayerRepository.java
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PlayerRepository extends JpaRepository<Player, Integer> {
+public interface PlayerRepository extends JpaRepository<Player, Long> {
 
     @EntityGraph(attributePaths = {"member", "room"})
-    Optional<Player> findByMember_Id(Integer memberId);
+    Optional<Player> findByMember_Id(Long memberId);
 
     @EntityGraph(attributePaths = {"member", "room"})
-    Optional<Player> findByMember_IdAndRoom_Id(Integer memberId, Integer roomId);
+    Optional<Player> findByMember_IdAndRoom_Id(Long memberId, Long roomId);
 
     @EntityGraph(attributePaths = {"member", "room"})
     List<Player> findAllByRoom(Room room);
